+++ MatrixSDK/Utils/MXHTTPOperation.h
@@ -31,7 +31,7 @@
  The underlying HTTP request.
  The reference changes in case of retries.
  */
-@property (nonatomic) NSURLSessionDataTask *operation;
+@property (nonatomic, nullable) NSURLSessionDataTask *operation;
 
 /**
  The age in milliseconds of the instance.
@@ -61,6 +61,11 @@
 @property (nonatomic, readonly, getter=isCancelled) BOOL canceled;
 
 /**
+ The HTTP response.
+ */
+@property (nonatomic, nullable) NSHTTPURLResponse *httpResponse;
+
+/**
  Cancel the HTTP request.
  */
 - (void)cancel;
@@ -74,7 +79,7 @@
  
  @param operation the other operation to copy data from. If the other operation is nil do nothing.
  */
-- (void)mutateTo:(MXHTTPOperation*)operation;
+- (void)mutateTo:(MXHTTPOperation* _Nullable)operation;
 
 /**
  Extract the NSHTTPURLResponse from an error.
@@ -82,6 +87,6 @@
  @param error the request error.
  @return the HTTP response.
  */
-+ (NSHTTPURLResponse *)urlResponseFromError:(NSError*)error;
++ (NSHTTPURLResponse * _Nullable)urlResponseFromError:(NSError* _Nullable)error;
 
 @end